requirements.txt
streamlit
torch
transformers
Pillow
soundfile
torchaudio
wikipedia
